starting to restructure state machine
